Add reference counting and state to channel group items.
Motivation: thc is cooperative, but there are still cases where
we can get into trouble. In thc_poll_recv, it's possible that
the awe we switch to (to receive a response) wants to kill the
channel that we just polled on. In the original implementation,
after we scheduled that awe, we would eventually come back into
thc_poll_recv, go to the top of the loop, and try to receive
again; but instead, we should just abort (and in fact, we would
be using memory after it had been freed).

When a channel group item is created (with thc_channel_group_item_init),
its state = LIVE, and its ref count is 1. So, the creating awe
automatically gets a reference. When another awe wants to grab
a reference, it should call thc_channel_group_item_inc_ref; and
to release a reference, thc_channel_group_item_dec_ref. dec ref
will remove the channel group item from its group and free it
when the ref count reaches 0. To signal to other awe's (thc async
threads) that a channel is dead, you should call
thc_channel_group_item_mark_dead.