Add thc_channel definition, update pieces of interface.
I realized we need to be doing the state checking in some of
the other ipc functions (perhaps indirectly, e.g. thc_ipc_call).
And then it seemed to make more sense to create a
new data type - a thc channel that wraps a fipc channel with
extras that we need when doing async stuff. So, we still do
ref counting, but it's on the thc_channel.

I no longer free the thc_channel when the ref drops. Leave up
to the libasync user to decide.

I updated the interface so that e.g. thc_ipc_recv now takes
a struct thc_channel instead.