Add THCDumpAllStacks.
This will not work unless you implement thc_dump_stack_0 and
thc_link_to_frame_0 for your architecture. Moreover, the implementation
of thc_dump_stack_0 should continue walking the stack until it
hits either a null frame or return address. (The usual way is to
assume stacks are aligned at certain addresses and to stop the walking
when you reach that alignment. That WILL NOT work here.)