ptstate: Add additional members for handling sync EPs
For all the callbacks registered with the kernel, the invocation can be from a
kernel thread, user thread or from KLCD itself. If it's from KLCD, there is
nothing that needs to be setup. If it's from a non LCD context, we need to
setup the runtime everytime. For now, grant the sync endpoint and don't delete
the cptr, because we don't exactly know when the process would exit.

Signed-off-by: Vikram Narayanan <vikram186@gmail.com>